feat: Use the babelrc dirname to set the working directory (#71)
Fixes #61

Now, the paths inside the configuration must be relative to the babelrc file instead of the working directory, because sometimes, the working directory changes and is not clear for the user. By forcing them to be relative to the babelrc file, I hope to resolve most of the issues with ava and others systems.

BREAKING CHANGE: The root paths and alias in the plugin configuration must be relative to the babelrc file itself instead of the working directory.