Replace Babel 6 with Babel 7 (`@babel/core`) in tests
This commit adds coverage for Babel 7, namely plugins using the `@babel` scope. Also, the plugin format for Babel 7 has changed -- namely, Babel 6 used to pass in `"unknown"` as the filename for anonymous code snippets, which is no longer the case. To preserve backwards compatibility and to align with Babel's documentation that says the `filename` option is optional, this plugin simulates Babel 6's behavior when the filename is undefined.

Test plan: Updated unit tests and verified all pass without warnings. No lint warnings/errors.